+++ CHANGELOG.md
@@ -134,6 +134,7 @@
   - `$HOME`, `/tmp`, `/var/tmp` are bind mounted from the host.
   - The full `/dev` is bind mounted from the host, unless `mount dev = minimal`
     in `singularity.conf`.
+  - `bind path` entries in `singularity.conf` are mounted into the container.
 
 ### Developer / API